requirements.txt
# API Management Service Requirements
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
httpx>=0.24.0
psycopg2-binary>=2.9.9
uuid>=1.30 